feat: api/player.php に音量初期値リセット・字幕補正を追加、nowplaying に現在キー・プレイヤー種別を追加
モバイルアプリのリモコン画面 Web 版同等化のためのサーバー側対応:

- mpcctrl_bs5.php 冒頭の音量初期値リセット・字幕補正の実装を
  function_playeradjust.php に切り出し (Web 版の挙動は不変)
- api/player.php に action を追加:
  - volume_reset: 曲開始時の初期音量に戻す (data.volume で新値を返す)
  - comp_get / comp_up / comp_down / comp_reset: 字幕補正 (data.comp_level)
- api/nowplaying.php (function_playingstatus.php) に追加:
  - player: プレイヤー種別 mpc|foobar|none (バッジ表示用)
  - keychange: 再生中の曲の現在キー (キー変更 UI 用)
- api/README.md を更新
